Dry up resize method across BigArray implementations
This method is the exact same across all implementations.
