+++ server/src/main/java/org/elasticsearch/common/util/AbstractBigByteArray.java
@@ -8,7 +8,11 @@
 
 package org.elasticsearch.common.util;
 
+import org.apache.lucene.util.ArrayUtil;
+import org.apache.lucene.util.RamUsageEstimator;
 import org.elasticsearch.common.recycler.Recycler;
+
+import java.util.Arrays;
 
 abstract class AbstractBigByteArray extends AbstractBigArray {
 
@@ -21,6 +25,23 @@ abstract class AbstractBigByteArray extends AbstractBigArray {
         for (int i = 0; i < pages.length; ++i) {
             pages[i] = newBytePage(i);
         }
+    }
+
+    /** Change the size of this array. Content between indexes <code>0</code> and <code>min(size(), newSize)</code> will be preserved. */
+    @Override
+    public void resize(long newSize) {
+        final int numPages = numPages(newSize);
+        if (numPages > pages.length) {
+            pages = Arrays.copyOf(pages, ArrayUtil.oversize(numPages, RamUsageEstimator.NUM_BYTES_OBJECT_REF));
+        }
+        for (int i = numPages - 1; i >= 0 && pages[i] == null; --i) {
+            pages[i] = newBytePage(i);
+        }
+        for (int i = numPages; i < pages.length && pages[i] != null; ++i) {
+            pages[i] = null;
+            releasePage(i);
+        }
+        this.size = newSize;
     }
 
     protected final byte[] newBytePage(int page) {
